refactor : 주석 수정 + 유저 검증 추가

+++ src/main/java/com/sparta/restplaceforj/service/PostService.java
@@ -6,6 +6,7 @@ import com.sparta.restplaceforj.dto.PostRequestDto;
 import com.sparta.restplaceforj.dto.PostResponseDto;
 import com.sparta.restplaceforj.entity.Post;
 import com.sparta.restplaceforj.entity.ThemeEnum;
+import com.sparta.restplaceforj.entity.User;
 import com.sparta.restplaceforj.exception.CommonException;
 import com.sparta.restplaceforj.exception.ErrorEnum;
 import com.sparta.restplaceforj.repository.PostDslRepository;
@@ -31,10 +32,14 @@ public class PostService {
   private final PostDslRepository postDslRepository;
 
   /**
-   * 글 새성.
+   * 글 생성
+   *
+   * @param postRequestDto : title, content, address, theme, placeName
+   * @param user           : 토큰 유저 정보
+   * @return PostResponseDto : title, content, address, likesCount, viewCount, themeEnum
    */
   @Transactional
-  public PostResponseDto createPost(PostRequestDto postRequestDto) {
+  public PostResponseDto createPost(PostRequestDto postRequestDto, User user) {
 
     try {
       ThemeEnum.valueOf(postRequestDto.getTheme());
@@ -45,6 +50,7 @@ public class PostService {
     Post post = postRepository.save(
         Post.builder()
             .requestDto(postRequestDto)
+            .user(user)
             .build());
 
     Post savedPost = postRepository.save(post);
@@ -54,13 +60,23 @@ public class PostService {
         .build();
   }
 
+  /**
+   * 삭제
+   *
+   * @param postId pathVariable
+   * @param user   : 토큰 유저 정보
+   */
   @Transactional
-  public void deletePost(long postId) {
+  public void deletePost(long postId, User user) {
+    Post post = postRepository.findByIdOrThrow(postId);
+    if (post.getUser().getId() != user.getId()) {
+      throw new CommonException(ErrorEnum.POST_MISMATCH);
+    }
+
     postRepository.deleteById(postId);
   }
 
-  public PageResponseDto getPlaceList(
-      int page, int size, String region, String theme) {
+  public PageResponseDto<String> getPlaceList(int page, int size, String region, String theme) {
 
     ThemeEnum themeEnum = ThemeEnum.valueOf(theme);
     Pageable pageRequest = PageRequest.of(page, size);
@@ -74,9 +90,16 @@ public class PostService {
   }
 
   /**
-   * 글 아이디와 제목만 조회.
+   * 글을 조회하면 제목과 아이디만 반환  .
+   *
+   * @param page      현재 페이지
+   * @param size      페이지 크기
+   * @param placeName 장소명
+   * @param q         검색질문
+   * @param sortBy    정렬 기준
+   * @return PageResponseDto : placeNameList, size, page, totalPages, totalElements
    */
-  public PageResponseDto getPostTitleList(
+  public PageResponseDto<PostIdTitleDto> getPostTitleList(
       int page, int size, String placeName, String sortBy, String q) {
 
     if (!(sortBy.equals("createAt") || sortBy.equals("viewsCount") ||
@@ -97,11 +120,20 @@ public class PostService {
   }
 
   /**
-   * 글 수정.
+   * 글 수정
+   *
+   * @param postId         수정 글 아이디
+   * @param postRequestDto title, content, address, theme, placeName
+   * @param user           : 토큰 유저 정보
+   * @return PostResponseDto id, userId, title, content, address, likesCount, viewsCount, themeEnum
    */
   @Transactional
-  public PostResponseDto updatePost(long postId, PostRequestDto postRequestDto) {
+  public PostResponseDto updatePost(long postId, PostRequestDto postRequestDto, User user) {
     Post post = postRepository.findByIdOrThrow(postId);
+    if (post.getUser().getId() != user.getId()) {
+      throw new CommonException(ErrorEnum.POST_MISMATCH);
+    }
+
     post.update(postRequestDto);
     return PostResponseDto.builder()
         .post(post)
@@ -109,10 +141,15 @@ public class PostService {
   }
 
   /**
-   * 글 단권 조회.
+   * 단권글 조회.
+   *
+   * @param postId 조회 글 아이디
+   * @return PostResponseDto :id, userId, title, content, address, likesCount, viewsCount, themeEnum
    */
+  @Transactional
   public PostResponseDto getPost(long postId) {
     Post post = postRepository.findByIdOrThrow(postId);
+    post.addViewToPost();
 
     return PostResponseDto.builder()
         .post(post)
